Added <, >, <=, and >= operators to Vector2. Finished and tested AnimatedComponent. Removed events from Scene/Layer for update/render. Fixed .size() bug in ArrayList's merge function. Fixed clone() function in FrameData. Fixed formatting and parameter ordering on all Components. Updated testing project.

+++ Arc Testing/SpinMenu.cpp
@@ -32,7 +32,7 @@ void SpinMenu::update( const FrameData* data )
 
 	if (_trans)
 	{
-		_transSpeed -= data->getElapsedMilliseconds();
+		_transSpeed -= (float)data->getElapsedMilliseconds();
 		_transAmount = clamp(_transSpeed / _transSpeedMax, 0.0f, 1.0f);
 
 		if (_transDir == Direction::DIR_SOUTH)
